add docstrings
